STYLE: Use the episode order feature
Use the episode order feature.

Using this feature allows to override the ordering determined by the episode
filenames. This is also useful to avoid needing to rename all subsequent files
to keep their numbers consistent when a new episode has to be added somewhere in
the middle.

Following from the above:
- Remove the numbers from the filenames and folders.
- Rename the figure folders to match their corresponding episode names.

Take advantage of the commit to fix the figure relative location in some of the
Jupyter notebooks.