ENH: Reuse the anatomical orientation views generating method
Reuse the anatomical orientation views generating method in the deterministic
and probabilistic tractography episodes.

Put the method in a module dwelling in a `utils` package so that the module can
be easily imported.

Re-order the scenes so that their sequence matches the existing pure
`matplotlib`-based figures.

Take advantage of the commit to split the output image cells from the next block
cells in the probabilistic episode Jupyter notebook.

Set to `None` the default value of the colormap in the anatomical orientation
views generating method in order to avoid applying an undesired colormap to the
tractograms.