BUG: Fix peak visualization method call and limit scope
Fix the call to the peak visualization helper method and limit the scope of the
imports.

Importing the visualization helper methods at the beginning causes the tests to
fail due to the fact that the `utils` package is not being provided, causing
```
Notebook cell execution failed
Cell 0: Cell execution caused an exception

Traceback:

---------------------------------------------------------------------------
ModuleNotFoundError                       Traceback (most recent call last)
<ipython-input-1-34b4a4d05158> in <module>
     11 from fury import actor, colormap
     12
---> 13 from utils.visualization_utils import generate_anatomical_volume_figure
     14
     15

ModuleNotFoundError: No module named 'utils'
```

when testing the notebooks.

Limiting the scope of the imports to the cells that are skipped serves as a
workaround.

The direct imports from `fury` are also moved for the sake of consistency.